+++ st-testcase-uniform/predefines.h
@@ -100,7 +100,7 @@
  * Files: predefines.h, problem-definition.h, logger.h are assumed to be compatible with this API version
  * For list of API versions see: https://gitlab.fizyka.pw.edu.pl/wtools/wslda/-/wikis/API-version
  * */
-#define API_VERSION 20230426
+#define API_VERSION 20231218
 
 /**
  * activate this flag for setting code in testing mode with uniform system