+++ st-project-template/predefines.h
@@ -101,5 +101,5 @@
  * Files: predefines.h, problem-definition.h, logger.h are assumed to be compatible with this API version
  * For list of API versions see: https://gitlab.fizyka.pw.edu.pl/wtools/wslda/-/wikis/API-version
  * */
-#define API_VERSION 20231218
+#define API_VERSION 20240131