+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 2023.xx.xx
 * Minor improvements of inittype=0 mode.
 * New API_VERSION=20231218. User can define new wdata variables via logger.h.
+* New API_VERSION=20240131. User can write wave functions from td code to file via logger.h.
 
 # 2023.09.16
 * New API_VERSION=20230426. Added modify_densities(...) in td codes.